+++ R/article.R
@@ -244,6 +244,17 @@ jasa_article <- function(
   )
 }
 
+#' @section `josis_article`: Format for creating submissions to the
+#' Journal of Spatial Information Science. Adapted from
+#' \samp{https://www.overleaf.com/latex/templates/journal-of-spatial-information-science-template/bmdhbgxnhtqx}.
+#' @export
+#' @rdname article
+josis_article <- function(..., keep_tex = TRUE, citation_package = 'default') {
+  pdf_document_format(
+    "josis", keep_tex = keep_tex, citation_package = citation_package, ...
+  )
+}
+
 #' @section `lipics_article`: Format for creating submissions to
 #'   LIPIcs - Leibniz International Proceedings Informatics - articles.
 #'   Adapted from the official Instructions for Authors at